+++ Vault-Coder/src/utopia/vault/coder/model/data/ClassReferences.scala
@@ -7,4 +7,5 @@ import utopia.vault.coder.model.scala.datatype.Reference
   * @author Mikko Hilpinen
   * @since 14.10.2021, v1.2
   */
-case class ClassReferences(model: Reference, data: Reference, factory: Reference, dbModel: Reference)
+case class ClassReferences(model: Reference, data: Reference, factory: Reference, dbModel: Reference,
+                           genericAccessTrait: Option[Reference])